+++ src/app/admin/loading.tsx
@@ -1,0 +1,7 @@
+export default function AdminLoading() {
+  return (
+    <div className="flex items-center justify-center py-12">
+      <div className="h-8 w-8 animate-spin rounded-full border-4 border-gray-300 border-t-gray-600" />
+    </div>
+  );
+}
